demos/23.05.24/requirements.txt
certifi==2023.5.7
charset-normalizer==3.1.0
climage==0.1.3
idna==3.4
kdtree==0.16
mypy==1.3.0
mypy-extensions==1.0.0
Pillow==9.5.0
python-dotenv==1.0.0
requests==2.31.0
typing_extensions==4.6.1
urllib3==2.0.2
